feat: integrate LOBOROBOT hardware driver into navigation pipeline
- add src/hardware/ package with loborobot.py (PCA9685 + LOBOROBOT + Robot singleton)
  - fixed FORWARD_SPEED=30, ROTATE_SPEED=30
  - calibrated V_FORWARD=0.215 m/s, V_ROTATE=96.0 deg/s (measured)
  - logging on each motion method (action, distance/angle, duration)
- car_control.py: wire CarController to src.hardware.Robot
  - add UNIT_TO_METERS=0.185 constant for grid-unit → meter conversion
  - hardware present → real motor calls; absent → mock sleep fallback
- navigation.py: import HARDWARE_AVAILABLE from src.hardware
  - pause_navigation / stop_navigation call t_stop() immediately on hardware
  - status endpoint exposes hardware_available field
- parser.py: remove int() truncation on dx/dy to support float coordinates
  - distances type: list[int] → list[float]
  - get_absolute_direction signature: int → float
- typedef.py: CarAction.distance int → float (unit: map grid, converted downstream)

+++ backend/src/smart_triager/car/parser.py
@@ -53,7 +53,7 @@ def expand_main_route_to_full_path(
     return full_path
 
 
-def get_absolute_direction(dx: int, dy: int) -> str:
+def get_absolute_direction(dx: float, dy: float) -> str:
     """
     根据坐标差计算绝对方向
 
@@ -172,7 +172,7 @@ def parse_route_to_commands(
 
     # 2. 计算绝对方向序列和距离
     directions: list[str] = []
-    distances: list[int] = []
+    distances: list[float] = []
 
     for i in range(len(full_path) - 1):
         node_id1 = full_path[i]
@@ -184,11 +184,11 @@ def parse_route_to_commands(
         if not node1 or not node2:
             raise ValueError(f"Node not found: {node_id1} or {node_id2}")
 
-        dx = int(node2.x - node1.x)  # 坐标转换为整数
-        dy = int(node2.y - node1.y)
+        dx = node2.x - node1.x  # 坐标差（米）
+        dy = node2.y - node1.y
 
         direction = get_absolute_direction(dx, dy)
-        distance = abs(dx) + abs(dy)  # 曼哈顿距离
+        distance = abs(dx) + abs(dy)  # 曼哈顿距离（米）
 
         directions.append(direction)
         distances.append(distance)
